authorization

+++ task-manager-typeorm/src/tasks/tasks.service.ts
@@ -1,3 +1,4 @@
+import { User } from './../auth/user.entity';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTaskFilterDto } from './dto/get-filtered-tasks.dto';
@@ -13,12 +14,14 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTaskFilterDto, user:User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto,user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+    async getTaskById( id: number, user ): Promise<Task> {
+
+    // we are quering ID with respect to user
+    const found = await this.taskRepository.findOne( {where:{id,userId:user.id}} );
 
     if (!found) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -27,28 +30,20 @@ export class TasksService {
     return found;
   }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        const { title, description } = createTaskDto;
-
-        const task = new Task();
-        task.title = title;
-        task.description = description;
-        task.status = TaskStatus.OPEN;
-        await task.save();
-
-        return task;
+    async createTask(createTaskDto: CreateTaskDto, user:User): Promise<Task> {
+        return this.taskRepository.createTask(createTaskDto,user);
     }
 
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user:User): Promise<void> {
+    const result = await this.taskRepository.delete({id,userId:user.id});
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID "${id}" not found`);
     }
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user:User): Promise<Task> {
+    const task = await this.getTaskById(id,user);
     task.status = status;
     await task.save();
     return task;
